revert merge eventsTable
