+++ debian/changelog
@@ -1,8 +1,8 @@
 wb-scenarios (1.10.1) stable; urgency=medium
 
-  * Merge action/event registries into control-interaction-registry.mod.js
+  * Merge the two action tables into table-handling-actions.mod.js
 
- -- Valerii Trofimov <valeriy.trofimov@wirenboard.com>  Tue, 21 Jul 2026 18:00:00 +0300
+ -- Valerii Trofimov <valeriy.trofimov@wirenboard.com>  Wed, 22 Jul 2026 15:18:00 +0300
 
 wb-scenarios (1.10.0) stable; urgency=medium
