+++ scenarios/astronomical-timer/astronomical-timer.mod.js
@@ -11,7 +11,7 @@ var ScenarioBase = require('wbsc-scenario-base.mod').ScenarioBase;
 var ScenarioState = require('virtual-device-helpers.mod').ScenarioState;
 var Logger = require('logger.mod').Logger;
 var SunCalc = require('suncalc.mod');
-var registry = require('control-interaction-registry.mod');
+var registry = require('table-handling-actions.mod');
 var constants = require('constants.mod');
 var durationReverse = require('duration-reverse.mod');
 var isControlTypeValid =
